\name{confounding-package}
\alias{confounding-package}
\alias{confounding}
\docType{package}
\title{
Analyzing Democratic Peace Survey for "Confounding in Survey Experiments"
}
\description{
Code and data to analyze Democratic Peace Survey
\details{
\tabular{ll}{
Package: \tab confounding_survey\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-06-30\cr
License: \tab MIT\cr
}

}

\author{
Baobao Zhang <baobao.zhang@yale.edu>
}

\references{
Dafoe, Allan, Baobao Zhang, and Devin Caughey. "Confounding in Survey Experiment" 2015. Working Paper.
}

\keyword{ survey }
\seealso{

}
\examples{

}
