\name{confounding_survey-package}
\alias{confounding_survey-package}
\alias{confounding_survey}
\docType{package}
\title{
Analyzing Democratic Peace Survey
}
\description{
Code and data to analyze Democratic Peace Survey
\details{
\tabular{ll}{
Package: \tab confounding_survey\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-06-30\cr
License: \tab What license is it under?\cr
}

}
\author{
Baobao Zhang <baobao.zhang@yale.edu>
}
\references{

}

\keyword{ survey }
\seealso{

}
\examples{

}
